PDJB-NONE: Extract row factories for gas safety and electrical CYA step configs (#1259)
